refactor(frontend): own guest member token result in Feed Flow
